requirements.txt
mushroom-cli